% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{run_query}
\alias{run_query}
\title{Run query on presto}
\usage{
run_query(conn, sql_query)
}
\arguments{
\item{conn}{Connection object for Presto server}

\item{sql_query}{Query to run}
}
\value{
A dataframe with the query result
}
\description{
Run query on presto
}
\examples{
run_query(conn,"select * from my_table limit 1")
}
