% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{start_connection}
\alias{start_connection}
\title{Start connection}
\usage{
start_connection(host, port, user, schema, catalog, source)
}
\arguments{
\item{host}{Presto server's addres}

\item{port}{Presto server's port}

\item{port}{Presto server's user for making queries}

\item{port}{Presto server's schema}

\item{port}{Presto server's catalog}

\item{port}{Presto server's source}
}
\description{
Start connection
}
